Move SardanaBuffer to its own module
